Revert "merge simplify loading state dict #7849"
This reverts commit facc18caaec472cd46d080858cab68e3337c8fcd.
